+++ backend/canisters/community/impl/src/updates/toggle_mute_notifications.rs
@@ -2,8 +2,6 @@ use crate::{model::channels::MuteChannelResult, mutate_state, run_regular_jobs, 
 use canister_tracing_macros::trace;
 use community_canister::toggle_mute_notifications::{Response::*, *};
 use ic_cdk_macros::update;
-use msgpack::serialize_then_unwrap;
-use types::Empty;
 
 #[update]
 #[trace]
@@ -47,12 +45,8 @@ fn toggle_mute_notifications_impl(args: Args, state: &mut RuntimeState) -> Respo
             };
 
             if updated {
-                let user_canister_id = member.user_id.into();
-                state.data.fire_and_forget_handler.send(
-                    user_canister_id,
-                    "c2c_mark_community_updated_for_user_msgpack".to_string(),
-                    serialize_then_unwrap(Empty {}),
-                );
+                let user_id = member.user_id;
+                state.data.mark_community_updated_in_user_canister(user_id);
             }
             Success
         }