+++ backend/canisters/group/api/src/updates/c2c_tip_message.rs
@@ -1,13 +1,17 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{CompletedCryptoTransaction, MessageId, MessageIndex, UserId};
+use types::{CanisterId, Cryptocurrency, MessageId, MessageIndex, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
-    pub message_sender: UserId,
+    pub recipient: UserId,
     pub thread_root_message_index: Option<MessageIndex>,
     pub message_id: MessageId,
-    pub transfer: CompletedCryptoTransaction,
+    pub ledger: CanisterId,
+    pub token: Cryptocurrency,
+    pub amount: u128,
+    pub username: String,
+    pub display_name: Option<String>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -15,7 +19,7 @@ pub enum Response {
     Success,
     MessageNotFound,
     CannotTipSelf,
-    MessageSenderMismatch,
+    RecipientMismatch,
     NotAuthorized,
     GroupFrozen,
     UserNotInGroup,