+++ backend/canisters/user/impl/src/updates/tip_message.rs
@@ -1,10 +1,11 @@
 use crate::crypto::process_transaction;
 use crate::guards::caller_is_owner;
 use crate::{mutate_state, read_state, run_regular_jobs, RuntimeState};
+use candid::Principal;
 use canister_tracing_macros::trace;
-use chat_events::TipMessageResult;
+use chat_events::{TipMessageArgs, TipMessageResult};
 use ic_cdk_macros::update;
-use types::{Chat, ChatId, CompletedCryptoTransaction, EventIndex, MessageId, MessageIndex, UserId};
+use types::{icrc1, Chat, ChatId, CommunityId, EventIndex, PendingCryptoTransaction, TimestampNanos, UserId};
 use user_canister::tip_message::{Response::*, *};
 
 #[update(guard = "caller_is_owner")]
@@ -12,124 +13,143 @@ use user_canister::tip_message::{Response::*, *};
 async fn tip_message(args: Args) -> Response {
     run_regular_jobs();
 
-    let message_sender = match read_state(|state| prepare(&args, state)) {
-        Ok(user_id) => user_id,
+    let (prepare_result, now_nanos) = match read_state(|state| prepare(&args, state)) {
+        Ok(ok) => ok,
         Err(response) => return *response,
     };
 
+    let pending_transfer = PendingCryptoTransaction::ICRC1(icrc1::PendingCryptoTransaction {
+        ledger: args.ledger,
+        token: args.token.clone(),
+        amount: args.amount,
+        to: Principal::from(args.recipient).into(),
+        fee: args.fee,
+        memo: None,
+        created: now_nanos,
+    });
     // Make the crypto transfer
-    let transfer = match process_transaction(args.transfer).await {
-        Ok(completed) => completed,
+    let completed_transfer = match process_transaction(pending_transfer).await {
+        Ok(transfer) => transfer,
         Err(failed) => return TransferFailed(failed.error_message().to_string()),
     };
 
-    match args.chat {
-        Chat::Direct(chat_id) => mutate_state(|state| {
-            tip_direct_chat_message(chat_id, args.thread_root_message_index, args.message_id, transfer, state)
-        }),
-        Chat::Group(chat_id) => {
+    match prepare_result {
+        PrepareResult::Direct(tip_message_args) => mutate_state(|state| tip_direct_chat_message(tip_message_args, state)),
+        PrepareResult::Group(group_id, c2c_args) => {
             use group_canister::c2c_tip_message::Response;
-            let args = group_canister::c2c_tip_message::Args {
-                message_sender,
-                thread_root_message_index: args.thread_root_message_index,
-                message_id: args.message_id,
-                transfer,
-            };
-            match group_canister_c2c_client::c2c_tip_message(chat_id.into(), &args).await {
+            match group_canister_c2c_client::c2c_tip_message(group_id.into(), &c2c_args).await {
                 Ok(Response::Success) => Success,
                 Ok(Response::MessageNotFound) => MessageNotFound,
                 Ok(Response::CannotTipSelf) => CannotTipSelf,
-                Ok(Response::MessageSenderMismatch) => TransferNotToMessageSender,
+                Ok(Response::RecipientMismatch) => TransferNotToMessageSender,
                 Ok(Response::NotAuthorized) => NotAuthorized,
                 Ok(Response::GroupFrozen) => ChatFrozen,
                 Ok(Response::UserNotInGroup) => ChatNotFound,
                 Ok(Response::UserSuspended) => UserSuspended,
-                Err(error) => InternalError(format!("{error:?}"), Box::new(args.transfer)),
+                Err(error) => InternalError(format!("{error:?}"), Box::new(completed_transfer)),
             }
         }
-        Chat::Channel(community_id, channel_id) => {
+        PrepareResult::Channel(community_id, c2c_args) => {
             use community_canister::c2c_tip_message::Response;
-            let args = community_canister::c2c_tip_message::Args {
-                message_sender,
-                channel_id,
-                thread_root_message_index: args.thread_root_message_index,
-                message_id: args.message_id,
-                transfer,
-            };
-            match community_canister_c2c_client::c2c_tip_message(community_id.into(), &args).await {
+            match community_canister_c2c_client::c2c_tip_message(community_id.into(), &c2c_args).await {
                 Ok(Response::Success) => Success,
                 Ok(Response::MessageNotFound) => MessageNotFound,
                 Ok(Response::CannotTipSelf) => CannotTipSelf,
-                Ok(Response::MessageSenderMismatch) => TransferNotToMessageSender,
+                Ok(Response::RecipientMismatch) => TransferNotToMessageSender,
                 Ok(Response::NotAuthorized) => NotAuthorized,
                 Ok(Response::CommunityFrozen) => ChatFrozen,
                 Ok(Response::UserSuspended) => UserSuspended,
                 Ok(Response::UserNotInCommunity | Response::ChannelNotFound) => ChatNotFound,
-                Err(error) => InternalError(format!("{error:?}"), Box::new(args.transfer)),
+                Err(error) => InternalError(format!("{error:?}"), Box::new(completed_transfer)),
             }
         }
     }
 }
 
-fn prepare(args: &Args, state: &RuntimeState) -> Result<UserId, Box<Response>> {
-    if args.transfer.is_zero() {
-        return Err(Box::new(TransferCannotBeZero));
-    }
+enum PrepareResult {
+    Direct(TipMessageArgs),
+    Group(ChatId, group_canister::c2c_tip_message::Args),
+    Channel(CommunityId, community_canister::c2c_tip_message::Args),
+}
 
-    let recipient = match args.transfer.user_id() {
-        Some(u) => u,
-        None => return Err(Box::new(TransferNotToMessageSender)),
-    };
-
+fn prepare(args: &Args, state: &RuntimeState) -> Result<(PrepareResult, TimestampNanos), Box<Response>> {
     let my_user_id: UserId = state.env.canister_id().into();
-    if my_user_id == recipient {
-        return Err(Box::new(CannotTipSelf));
-    }
-
-    if !match &args.chat {
-        Chat::Direct(d) => state.data.direct_chats.has(d),
-        Chat::Group(g) => state.data.group_chats.has(g),
-        Chat::Channel(c, _) => state.data.communities.has(c),
-    } {
-        return Err(Box::new(ChatNotFound));
-    }
-
     if state.data.suspended.value {
         Err(Box::new(UserSuspended))
-    } else if args.transfer.is_zero() {
+    } else if args.amount == 0 {
         Err(Box::new(TransferCannotBeZero))
+    } else if my_user_id == args.recipient {
+        Err(Box::new(CannotTipSelf))
     } else {
-        Ok(recipient)
+        let now_nanos = state.env.now_nanos();
+        match args.chat {
+            Chat::Direct(chat_id) if state.data.direct_chats.has(&chat_id) => Ok((
+                PrepareResult::Direct(TipMessageArgs {
+                    user_id: my_user_id,
+                    recipient: args.recipient,
+                    thread_root_message_index: args.thread_root_message_index,
+                    message_id: args.message_id,
+                    ledger: args.ledger,
+                    token: args.token.clone(),
+                    amount: args.amount,
+                    now: state.env.now(),
+                }),
+                now_nanos,
+            )),
+            Chat::Group(group_id) if state.data.group_chats.has(&group_id) => Ok((
+                PrepareResult::Group(
+                    group_id,
+                    group_canister::c2c_tip_message::Args {
+                        recipient: args.recipient,
+                        thread_root_message_index: args.thread_root_message_index,
+                        message_id: args.message_id,
+                        ledger: args.ledger,
+                        token: args.token.clone(),
+                        amount: args.amount,
+                        username: state.data.username.value.clone(),
+                        display_name: state.data.display_name.value.clone(),
+                    },
+                ),
+                now_nanos,
+            )),
+            Chat::Channel(community_id, channel_id) if state.data.communities.has(&community_id) => Ok((
+                PrepareResult::Channel(
+                    community_id,
+                    community_canister::c2c_tip_message::Args {
+                        recipient: args.recipient,
+                        channel_id,
+                        thread_root_message_index: args.thread_root_message_index,
+                        message_id: args.message_id,
+                        ledger: args.ledger,
+                        token: args.token.clone(),
+                        amount: args.amount,
+                        username: state.data.username.value.clone(),
+                        display_name: state.data.display_name.value.clone(),
+                    },
+                ),
+                now_nanos,
+            )),
+            _ => Err(Box::new(ChatNotFound)),
+        }
     }
 }
 
-fn tip_direct_chat_message(
-    chat_id: ChatId,
-    thread_root_message_index: Option<MessageIndex>,
-    message_id: MessageId,
-    transfer: CompletedCryptoTransaction,
-    state: &mut RuntimeState,
-) -> Response {
-    if let Some(chat) = state.data.direct_chats.get_mut(&chat_id) {
-        let my_user_id = state.env.canister_id().into();
-        let now = state.env.now();
-        match chat.events.tip_message(
-            my_user_id,
-            chat.them,
-            EventIndex::default(),
-            thread_root_message_index,
-            message_id,
-            transfer.clone(),
-            now,
-        ) {
+fn tip_direct_chat_message(args: TipMessageArgs, state: &mut RuntimeState) -> Response {
+    if let Some(chat) = state.data.direct_chats.get_mut(&args.recipient.into()) {
+        match chat.events.tip_message(args.clone(), EventIndex::default()) {
             TipMessageResult::Success => {
                 let c2c_args = user_canister::c2c_tip_message::Args {
-                    thread_root_message_index,
-                    message_id,
-                    transfer,
+                    thread_root_message_index: args.thread_root_message_index,
+                    message_id: args.message_id,
+                    ledger: args.ledger,
+                    token: args.token,
+                    amount: args.amount,
+                    username: state.data.username.value.clone(),
+                    display_name: state.data.display_name.value.clone(),
+                    user_avatar_id: state.data.avatar.value.as_ref().map(|a| a.id),
                 };
                 state.data.fire_and_forget_handler.send(
-                    chat_id.into(),
+                    args.recipient.into(),
                     "c2c_tip_message_msgpack".to_string(),
                     msgpack::serialize_then_unwrap(c2c_args),
                 );
@@ -137,7 +157,7 @@ fn tip_direct_chat_message(
             }
             TipMessageResult::MessageNotFound => MessageNotFound,
             TipMessageResult::CannotTipSelf => CannotTipSelf,
-            TipMessageResult::MessageSenderMismatch => TransferNotToMessageSender,
+            TipMessageResult::RecipientMismatch => TransferNotToMessageSender,
         }
     } else {
         ChatNotFound