+++ backend/libraries/types/src/notifications.rs
@@ -20,6 +20,9 @@ pub enum Notification {
     DirectReactionAdded(DirectReactionAddedNotification),
     GroupReactionAdded(GroupReactionAddedNotification),
     ChannelReactionAdded(ChannelReactionAddedNotification),
+    DirectMessageTipped(DirectMessageTipped),
+    GroupMessageTipped(GroupMessageTipped),
+    ChannelMessageTipped(ChannelMessageTipped),
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
@@ -131,6 +134,49 @@ pub struct ChannelReactionAddedNotification {
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct DirectMessageTipped {
+    pub them: UserId,
+    pub thread_root_message_index: Option<MessageIndex>,
+    pub message_index: MessageIndex,
+    pub message_event_index: EventIndex,
+    pub username: String,
+    pub display_name: Option<String>,
+    pub tip: String, // formatted amount, eg. "0.1 CHAT"
+    pub user_avatar_id: Option<u128>,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct GroupMessageTipped {
+    pub chat_id: ChatId,
+    pub thread_root_message_index: Option<MessageIndex>,
+    pub message_index: MessageIndex,
+    pub message_event_index: EventIndex,
+    pub group_name: String,
+    pub tipped_by: UserId,
+    pub tipped_by_name: String,
+    pub tipped_by_display_name: Option<String>,
+    pub tip: String,
+    pub group_avatar_id: Option<u128>,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct ChannelMessageTipped {
+    pub community_id: CommunityId,
+    pub channel_id: ChannelId,
+    pub thread_root_message_index: Option<MessageIndex>,
+    pub message_index: MessageIndex,
+    pub message_event_index: EventIndex,
+    pub community_name: String,
+    pub channel_name: String,
+    pub tipped_by: UserId,
+    pub tipped_by_name: String,
+    pub tipped_by_display_name: Option<String>,
+    pub tip: String,
+    pub community_avatar_id: Option<u128>,
+    pub channel_avatar_id: Option<u128>,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct CryptoTransferDetails {
     pub recipient: UserId,
     pub recipient_username: Option<String>,
@@ -177,5 +223,5 @@ fn notification_length() {
 
     let bytes = candid::encode_one(notification).unwrap().len();
 
-    assert!(bytes < 630, "{bytes}");
+    assert!(bytes < 850, "{bytes}");
 }