+++ frontend/app/src/stores/xframe.ts
@@ -5,12 +5,14 @@ import { setModifiedTheme } from "../theme/themes";
 import { routerReady } from "../routes";
 
 const FRAME_ANCESTORS = [
-    "https://ht7v7-iaaaa-aaaak-qakga-cai.icp0.io",
-    "https://mdocx-gyaaa-aaaak-qcbsq-cai.icp0.io",
-    "https://calm-pasca-49d7be.netlify.app",
     "http://localhost:5173",
-    "https://zexzi-jyaaa-aaaam-abj3q-cai.icp0.io",
-    "https://xw4dq-4yaaa-aaaam-abeuq-cai.ic0.app",
+    "https://calm-pasca-49d7be.netlify.app", // betbase proof of concept
+    "https://221bravo.app",
+    "https://ht7v7-iaaaa-aaaak-qakga-cai.icp0.io", //221bravo
+    "https://mdocx-gyaaa-aaaak-qcbsq-cai.icp0.io",
+    "https://zexzi-jyaaa-aaaam-abj3q-cai.icp0.io", //Betbase dev
+    "https://xw4dq-4yaaa-aaaam-abeuq-cai.ic0.app", //Betbase live
+    "https://spyzr-gqaaa-aaaan-qd66q-cai.icp0.io", // vaultbet
 ];
 
 type XFrameMessage = UpdateTheme | ChangeRoute;