+++ frontend/openchat-agent/src/services/community/candid/idl.js
@@ -1241,7 +1241,6 @@ export const idlFactory = ({ IDL }) => {
       'timestamp' : TimestampMillis,
       'pinned_messages' : IDL.Vec(MessageIndex),
       'latest_event_index' : EventIndex,
-      'rules' : Rules,
     }),
     'PrivateCommunity' : IDL.Null,
     'PrivateChannel' : IDL.Null,
@@ -1260,7 +1259,6 @@ export const idlFactory = ({ IDL }) => {
     'members_removed' : IDL.Vec(UserId),
     'timestamp' : TimestampMillis,
     'latest_event_index' : EventIndex,
-    'rules' : IDL.Opt(Rules),
     'blocked_users_added' : IDL.Vec(UserId),
   });
   const SelectedChannelUpdatesResponse = IDL.Variant({
@@ -1299,7 +1297,6 @@ export const idlFactory = ({ IDL }) => {
     'user_groups' : IDL.Vec(UserGroupDetails),
     'timestamp' : TimestampMillis,
     'latest_event_index' : EventIndex,
-    'rules' : Rules,
   });
   const SelectedInitialResponse = IDL.Variant({
     'Success' : SelectedInitialSuccess,
@@ -1318,7 +1315,6 @@ export const idlFactory = ({ IDL }) => {
     'user_groups' : IDL.Vec(UserGroupDetails),
     'members_removed' : IDL.Vec(UserId),
     'timestamp' : TimestampMillis,
-    'rules' : IDL.Opt(Rules),
     'blocked_users_added' : IDL.Vec(UserId),
   });
   const SelectedUpdatesResponse = IDL.Variant({