+++ frontend/openchat-agent/src/services/groupIndex/candid/types.d.ts
@@ -1211,7 +1211,6 @@ export interface SelectedGroupUpdates {
   'members_removed' : Array<UserId>,
   'timestamp' : TimestampMillis,
   'latest_event_index' : EventIndex,
-  'rules' : [] | [Rules],
   'blocked_users_added' : Array<UserId>,
 }
 export interface SetCommunityModerationFlagsArgs {