+++ frontend/openchat-agent/src/services/userIndex/candid/types.d.ts
@@ -1194,7 +1194,6 @@ export interface SelectedGroupUpdates {
   'members_removed' : Array<UserId>,
   'timestamp' : TimestampMillis,
   'latest_event_index' : EventIndex,
-  'rules' : [] | [Rules],
   'blocked_users_added' : Array<UserId>,
 }
 export interface SetDisplayNameArgs { 'display_name' : [] | [string] }